requirements.txt
streamlit>=1.25
PyPDF2>=3.0
faiss-cpu==1.7.4
sentence-transformers>=2.2.2
requests>=2.31
python-dotenv
tqdm
